feat: implement image upload functionality and refactor profile data handling; add new supabase clients for database and auth operations
