+++ clients/supabase/index.ts
@@ -28,123 +28,3 @@ AppState.addEventListener('change', (state) => {
     supabase.auth.stopAutoRefresh();
   }
 });
-
-export const supabaseClient = {
-  signInWithEmail: async (email: string, password: string) => {
-    const { error } = await supabase.auth.signInWithPassword({
-      email,
-      password,
-    });
-    if (error) {
-      throw new Error(error.message);
-    }
-  },
-  signInWithPhone: async (phone: string) => {
-    const { error } = await supabase.auth.signInWithOtp({
-      phone,
-    });
-    if (error) {
-      throw new Error(error.message);
-    }
-  },
-
-  verifyPhone: async (phone: string, verificationCode: string) => {
-    const { error } = await supabase.auth.verifyOtp({
-      phone,
-      token: verificationCode,
-      type: 'sms',
-    });
-    if (error) {
-      throw new Error(error.message);
-    }
-  },
-
-  getSession: async () => {
-    try {
-      const {
-        data: { session },
-      } = await supabase.auth.getSession();
-      return session;
-    } catch (error) {
-      console.error(error);
-      throw error;
-    }
-  },
-
-  signUp: async (phone: string, password: string) => {
-    const { error } = await supabase.auth.signUp({
-      phone,
-      password,
-    });
-    if (error) {
-      throw new Error(error.message);
-    }
-  },
-
-  updateUser: async (email: string, password: string, phone: string) => {
-    const { error } = await supabase.auth.updateUser({
-      email,
-      password,
-      data: {
-        phone_verified: true,
-        phone,
-      },
-    });
-    if (error) {
-      throw new Error(error.message);
-    }
-  },
-
-  resendVerificationCode: async (phone: string) => {
-    const { error } = await supabase.auth.resend({
-      phone,
-      type: 'sms',
-    });
-    if (error) {
-      throw new Error(error.message);
-    }
-  },
-
-  signOut: async () => {
-    const { error } = await supabase.auth.signOut();
-    if (error) {
-      throw new Error(error.message);
-    }
-  },
-
-  getUser: async (userId: string) => {
-    const { data, error } = await supabase
-      .from('users')
-      .select('*')
-      .eq('id', userId)
-      .single();
-    if (error) {
-      throw new Error(error.message);
-    }
-    return data;
-  },
-
-  getUserAddresses: async (userId: string) => {
-    const { data, error } = await supabase
-      .from('addresses')
-      .select('*')
-      .eq('user_id', userId)
-      .eq('type', 'USER_ADDRESS');
-    if (error) {
-      throw new Error(error.message);
-    }
-    return data;
-  },
-
-  getUserProfile: async (userId: string) => {
-    const { data, error } = await supabase
-      .from('profiles')
-      .select('first_name, last_name, abn, role')
-      .eq('user_id', userId)
-      .single();
-    if (error) {
-      throw new Error(error.message);
-    }
-    return data;
-  },
-};
