+++ utils/imagePicker/index.ts
@@ -1,0 +1,25 @@
+import * as ImagePicker from 'expo-image-picker';
+import { Alert } from 'react-native';
+
+export const imagePicker = {
+  requestImagePermission: async () => {
+    const { status } = await ImagePicker.requestMediaLibraryPermissionsAsync();
+    if (status !== 'granted') {
+      Alert.alert(
+        'Permission Required',
+        'Sorry, we need camera roll permissions to upload images.',
+      );
+    }
+  },
+
+  pickImage: async () => {
+    const result = await ImagePicker.launchImageLibraryAsync({
+      mediaTypes: ImagePicker.MediaTypeOptions.Images,
+      allowsEditing: true,
+      aspect: [4, 3],
+      quality: 0.8,
+      base64: true,
+    });
+    return result;
+  },
+};
